ClusteringService/requirements.txt
grpcio
grpcio-tools
numpy
scipy
scikit-learn